gfortran: Less noise and more variables

+++ gfortran-feedstock/recipe/activate-gfortran.sh
@@ -37,7 +37,6 @@ function _tc_activation() {
   local newval
   local from
   local to
-  local which
   local pass
 
   if [ "${act_nature}" = "activate" ]; then
@@ -56,9 +55,9 @@ function _tc_activation() {
           thing=$(echo "${thing}" | sed "s,^\([^\,]*\)\,.*,\1,")
           ;;
         *)
-          newval=$(which ${CONDA_PREFIX}/bin/${tc_prefix}${thing} 2>/dev/null)
-          if [ -z "${newval}" -a "${pass}" = "check" ]; then
-            echo "ERROR: This cross-compiler package contains no program ${CONDA_PREFIX}/bin/${tc_prefix}${thing}"
+          newval="${CONDA_PREFIX}/bin/${tc_prefix}${thing}"
+          if [ ! -x "${newval}" -a "${pass}" = "check" ]; then
+            echo "ERROR: This cross-compiler package contains no program ${newval}"
             return 1
           fi
           ;;
@@ -82,23 +81,48 @@ function _tc_activation() {
   return 0
 }
 
-env > /tmp/old-env-$$.txt
+# When people are using conda-build, assume that adding rpath during build, and pointing at
+#    the host env's includes and libs is helpful default behavior
+if [ "${CONDA_BUILD:-0}" = "1" ]; then
+  FFLAGS_USED="@FFLAGS@ -I${PREFIX}/include -fdebug-prefix-map=\${SRC_DIR}=/usr/local/src/conda/\${PKG_NAME}-\${PKG_VERSION} -fdebug-prefix-map=\${PREFIX}=/usr/local/src/conda-prefix"
+else
+  FFLAGS_USED="@FFLAGS@"
+fi
+
+if [ "${CONDA_BUILD:-0}" = "1" ]; then
+  if [ -f /tmp/old-env-$$.txt ]; then
+    rm -f /tmp/old-env-$$.txt || true
+  fi
+  env > /tmp/old-env-$$.txt
+fi
+
 _tc_activation \
   activate host @CHOST@ @CHOST@- \
   gfortran \
-  "FFLAGS,${FFLAGS:-@FFLAGS@}" \
-  "FORTRANFLAGS,${FORTRANFLAGS:-@FFLAGS@}" \
-  "DEBUG_FFLAGS,${DEBUG_FFLAGS:-@FFLAGS@ @DEBUG_FFLAGS@}"
+  "FFLAGS,${FFLAGS:-${FFLAGS_USED}}" \
+  "FORTRANFLAGS,${FORTRANFLAGS:-${FFLAGS_USED}}" \
+  "DEBUG_FFLAGS,${FFLAGS:-${FFLAGS_USED} @DEBUG_FFLAGS@}" \
+  "DEBUG_FORTRANFLAGS,${FORTRANFLAGS:-${FFLAGS_USED} @DEBUG_FFLAGS@}" \
 
-# extra ones:
-export FC="$GFORTRAN"
-export F95="$GFORTRAN"
+# extra ones - have a dependency on the previous ones, so done after.
+_tc_activation \
+  activate host @CHOST@ @CHOST@- \
+  "FC,${FC:-${GFORTRAN}}" \
+  "F77,${F77:-${GFORTRAN}}" \
+  "F90,${F90:-${GFORTRAN}}" \
+  "F95,${F95:-${GFORTRAN}}"
 
 if [ $? -ne 0 ]; then
   echo "ERROR: $(_get_sourced_filename) failed, see above for details"
 else
-  env > /tmp/new-env-$$.txt
-  echo "INFO: $(_get_sourced_filename) made the following environmental changes:"
-  diff -U 0 -rN /tmp/old-env-$$.txt /tmp/new-env-$$.txt | tail -n +4 | grep "^-.*\|^+.*" | grep -v "CONDA_BACKUP_" | sort
+  if [ "${CONDA_BUILD:-0}" = "1" ]; then
+    if [ -f /tmp/new-env-$$.txt ]; then
+      rm -f /tmp/new-env-$$.txt || true
+    fi
+    env > /tmp/new-env-$$.txt
+
+    echo "INFO: $(_get_sourced_filename) made the following environmental changes:"
+    diff -U 0 -rN /tmp/old-env-$$.txt /tmp/new-env-$$.txt | tail -n +4 | grep "^-.*\|^+.*" | grep -v "CONDA_BACKUP_" | sort
+    rm -f /tmp/old-env-$$.txt /tmp/new-env-$$.txt || true
+  fi
 fi
-rm -f /tmp/old-env-$$.txt /tmp/new-env-$$.txt || true
